+++ test/Fixture/Methods/NoParameterWithNullableTypeDeclarationRule/Success/MethodInClassWithParameterWithoutTypeDeclaration.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\NoParameterWithNullableTypeDeclarationRule\Success;
+
+final class MethodInClassWithParameterWithoutTypeDeclaration
+{
+    public function foo($bar)
+    {
+        return $bar;
+    }
+}